+++ .gitignore
@@ -1,3 +1,4 @@
 build
 .cache
 inputs/token.txt
+.vscode
