--- .vscode/tasks.json
@@ -1,25 +1,0 @@
-{
-    // See https://go.microsoft.com/fwlink/?LinkId=733558
-    // for the documentation about the tasks.json format
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "build",
-            "type": "shell",
-            "command": "cmake .. && make",
-            "options": {
-                "cwd": "${workspaceFolder}/build"
-            },
-            "dependsOn" : "clean"
-        },
-        {
-            "label": "clean",
-            "type": "shell",
-            "command": "rm -rf *",
-            "options": {
-                "cwd": "${workspaceFolder}/build"
-            },
-        }
-
-    ]
-}
