--- Findqinfo.cmake
@@ -1,9 +1,0 @@
-include(FetchContent)
-
-FetchContent_Declare(
-    qinfo
-    GIT_REPOSITORY git@github.com:Munich-Quantum-Software-Stack/QInfo.git
-    GIT_TAG testing
-)
-
-FetchContent_MakeAvailable(qinfo)
