+++ cmake/Findqinfo.cmake
@@ -1,0 +1,12 @@
+include(FetchContent)
+
+FetchContent_Declare(
+    qinfo
+    GIT_REPOSITORY git@github.com:Munich-Quantum-Software-Stack/qinfo.git
+    GIT_TAG develop
+)
+
+FetchContent_MakeAvailable(qinfo)
+FetchContent_GetProperties(qinfo)
+
+set(QINFO_INCLUDE_DIRS "${qinfo_SOURCE_DIR}/include")
