+++ src/wmi/CMakeLists.txt
@@ -1,0 +1,4 @@
+find_package(cjson)
+
+add_subdirectory(real)
+add_subdirectory(sim)
