+++ src/wmi/sim/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(WMI_INCLUDE_DIR ${CMAKE_SOURCE_DIR}/include/wmi/sim)
+
+add_library(backend_wmi_sim SHARED)
+target_sources(backend_wmi_sim PRIVATE wmi_sim.c)
+target_include_directories(backend_wmi_sim PUBLIC ${WMI_INCLUDE_DIR} ${QDMI_INCLUDE_DIRS} ${QDMI_SRC_DIRS} ${QINFO_INCLUDE_DIRS} ${CJON_INCLUDE_DIRS})
+target_link_libraries(backend_wmi_sim PUBLIC qdmi cjson curl)
