+++ tests/CMakeLists.txt
@@ -1,7 +1,10 @@
-add_executable(IBMBackend IBMBackendTest.c)
-target_link_libraries(IBMBackend backend_ibm)
-add_test(IBMBackend IBMBackend)
+add_subdirectory(ibm)
+add_subdirectory(wmi)
 
-add_executable(WMIBackend WMIBackendTest.c)
-target_link_libraries(WMIBackend backend_wmi_sim)
-add_test(WMIBackend WMIBackend)
+#add_executable(IBMBackend IBMBackendTest.c)
+#target_link_libraries(IBMBackend backend_ibm)
+#add_test(IBMBackend IBMBackend)
+
+#add_executable(WMIBackend WMIBackendTest.c)
+#target_link_libraries(WMIBackend backend_wmi_sim)
+#add_test(WMIBackend WMIBackend)
