+++ tests/ibm/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_executable(IBMBackend IBMBackendTest.c)
+target_link_libraries(IBMBackend backend_ibm)
+add_test(IBMBackend IBMBackend)
+
+#add_executable(WMIBackend WMIBackendTest.c)
+#target_link_libraries(WMIBackend backend_wmi_sim)
+#add_test(WMIBackend WMIBackend)
