+++ tests/inputs/Bell_circuit.bc
