+++ tests/inputs/basic_circuit.bc
