+++ tests/inputs/circuit_excited.bc
