+++ tests/inputs/circuit_ground.bc
