+++ tests/inputs/circuit_hadamard_decomposed.bc
