+++ tests/wmi/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(WMIBackend WMIBackendTest.c)
+target_link_libraries(WMIBackend backend_wmi_sim)
+add_test(WMIBackend WMIBackend)
