+++ tests/wmi/test.sh
